#include <11.h>.c

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -1,6 +1,5 @@
 #include <stdio.h>
 #include <math.h>
-
 /**
  * main - Entry point for the largest prime factor program
  * Description:
@@ -10,23 +9,23 @@
  */
 int main(void)
 {
-	long long num = 612852475143;
-	long long largestPrimeFactor = -1;
+	long num = 612852475143;
+	long largestPrimeFactor = -1;
 
 	while (num % 2 == 0) {
 		largestPrimeFactor = 2;
 		num /= 2;
 	}
 
-	long long limit = (long long)sqrt((double)num);
-	long long factor = 3;
+	long limit = (long)sqrt((double)num);
+	long factor = 3;
 	while (factor <= limit)
 	{
 		while (num % factor == 0)
 		{
 			largestPrimeFactor = factor;
 			num /= factor;
-			limit = (long long)sqrt((double)num);
+			limit = (long)sqrt((double)num);
 		}
 		factor += 2;
 	}
@@ -37,6 +36,5 @@ int main(void)
 	}
 
 	printf("%lld\n", largestPrimeFactor);
-
 	return (0);
 }
